refactor: consolidate availability and booking utilities; remove unused files and update imports
