+++ web-client/src/components/BookingServiceModal.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { DayPicker } from "react-day-picker";
 import "react-day-picker/dist/style.css";
 import { format } from "date-fns";
-import { getAvailableSlots } from "@/utils/availabilityUtils";
-import { bookSlot } from "@/utils/bookingsUtils";
+import { getAvailableSlots } from "@/utils/availability";
+import { bookSlot } from "@/utils/bookings";
 import { getAvailableDates } from "@/utils/seller";
 import { toast } from "sonner";
 import type { Service } from "@/types/types";
