+++ web-client/src/utils/dbdao.ts
@@ -23,3 +23,15 @@ export const fetchServices = async (
     };
   }
 };
+
+/**
+ * Fetch rows from Supabase with a match filter
+ */
+export async function fetchTable<T = unknown>(
+  table: string,
+  match: Record<string, unknown>,
+): Promise<T[]> {
+  const { data, error } = await supabase.from(table).select("*").match(match);
+  if (error) throw new Error(error.message);
+  return data as T[];
+}
